+++ trunk/xlib/xapp.h
@@ -97,10 +97,10 @@ public:
   void SetAtomUiso(TSAtom& sa, double val);
   
   template <class atom_list> static void UnifyPAtomList(atom_list& alist) {
-    const int ac = alist.Count();
-    for( int i=0; i < ac; i++ )
+    const size_t ac = alist.Count();
+    for( size_t i=0; i < ac; i++ )
       alist[i]->SetTag(i);
-    for( int i=0; i < ac; i++ )
+    for( size_t i=0; i < ac; i++ )
       if( alist[i]->GetTag() != i || alist[i]->IsDeleted() )
         alist[i] = NULL;
     alist.Pack();
